Logic for allowing null to pass.
Since null is actually org.codehaus.groovy.runtime.NullObject, it is
necessary to specify whether null should be allowed to pass.

This handles the use case in which null is an acceptable value.

Therefore, the methodology is to either set the default behavior of
whether or not to allow nulls to pass at the constructor level, or
override it at a particular method level.

While in many, a lot, or even most situations, there may not be a time
when the default behavior needs to be overridden; I'd hate for that one
person that encounters the use case to not be able to use it.