Added CoerciveSanityChecker .
One of the derivative sanity checkers that this plugin is really about.

There is some WEIRD behavior with isNumber() and isDouble() .

For example:

"1D".isNumber() // False
"1D."isDouble() // True

But, I'm going to allow this because that's the point of this checker:
it follows the coercion rules of String. Dems da rules.