Refactor of SanityChecker and SanityCheckReport.
SanityChecker was simplified.

It no longer starts off with anything to check against; that must be
declared via the check() method before anything can be done. The
IllegalStateExceptions are only there to catch developer folly.

SanityCheckReport was also simplified.

Much like SanityChecker, it is no longer tied to anything in particular.
This fixed some weird behavior when it was used as a List of
SanityCheckReports inside of the SanityChecker.

Tests for SanityCheckReport pending.